+++ app/lib/types/componentsTypes/_layout.ts
@@ -1,7 +1,7 @@
 import type { BottomTabHeaderProps } from '@react-navigation/bottom-tabs';
 import type { NativeStackHeaderProps } from '@react-navigation/native-stack';
 
-export type TypeHeaderHome = NativeStackHeaderProps;
+export type TypeHeaderHome = NativeStackHeaderProps | BottomTabHeaderProps;
 
 export type TypeHeaderSub = (NativeStackHeaderProps | BottomTabHeaderProps) & {
   goBack?: string | boolean;
